requirements.txt.txt
faiss-cpu
numpy
pandas
sentence-transformers
fastapi
pydantic
langchain
python-dotenv
scikit-learn
requests